+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -148,7 +148,7 @@ const uiSettings = {
 
 export const config = {
   [GALLERY_TYPE]: {
-    scrollingElement: document.getElementsByClassName('editor-example')[0],
+    scrollingElement: typeof window !== 'undefined' && document.getElementsByClassName('editor-example')[0],
   },
   [IMAGE_TYPE]: {
     imageEditorWixSettings: {